requirements.txt
attrs
flask==3.0.3
folium
requests
pandas
numpy
websockets
sqlitecloud>=0.0.84
